feat: add the courseOffering wrapper function
The courseOffering wrapper function allows a certain course offering to
be looked up.

API types have also been added. Plus, instead of directly returning an
object of the same structure as received from the API, it is modified a
bit for clarity. For example, requiredReadingNotes and registrarNotes,
among other note related properties, are all placed under a "notes"
property (so they become notes.requiredReading, notes.registrar, and so
on).