refactor: clean up helper files and change build folder from dist to lib  (#6)
* build: change the build folder from dist to lib

This change was done to make it clear that this code is meant to be used
as a library by other code. Also, `npm pack` was tested and works.

* refactor: clean up api-types and utils

Clean up the code in src/api-types and src/utils. In src/api-types,
functions were made to generate functions to prevent code duplication.