feat: add academic calendar courses api (#11)
* docs: add a contribution section to the README

Explain how contributions can be made by submitting a pull request.

* refactor: rename a function name and test name

Rename courseOutline to courseOffering for the wrapper function for
better clarity, and make the courseOffering test case name clearer too.

* test: add a test for convertStringDaysToEnumDays

Add a new test case for the convertStringDaysToEnumDays function.

* feat: add a course wrapper function

Add a wrapper function to get info about a course. This info does not
depend on the course section, and also gives a list of course sections
for the course.

* test: add a test for the course wrapper function

Add a test for the new course wrapper function. Also, update the test to
pass in the course number as a string. This is because the course number
could be a string value such as "100w", so only ever accepting a number
for the course number will not work.

* test: rename the course wrapper function test file

Rename the file for testing the course wrapper function to end in
".test.ts" instead of ".ts", so Jest can find it.

* test: add github workflow for running jest tests

Make it so that Jest tests are run when making a pull request on or
pushing code to GitHub.

* feat: try making a course get its course sections

Try to see if a Course can be made to get its CourseOffering sections,
by making API requests.

* feat: make a course able to get its sections

Make it so that given a Course class instance, it is possible to request
data from the API to get a CourseOffering for a given section.

* test: remove unused imports from the course tests

Remove imports that were not used in the Jest tests for the course
wrapper function.

* test: make github jest tests workflow save results

Make it so that the test results of the Jest tests are saved for the
GitHub workflow.

* test: make github workflow do pull requests only

Make it so that the Jest test GitHub workflow only targets pull
requests.

* docs: update README with year and term for example

Make it so that the example code includes the year and term for the
courseOffering wrapper function.