docs: add readme and typedoc documentation (#13)
* docs: add typedoc docs folder

Add typedoc, and make it output documentation to the folder ./docs.

* docs: add README sections and documentation

Add more sections to the README, and add documentation explaining how to
use wrapper functions and the custom classes.

* docs: center top README blocks and add a doc link

Make the blocks like "Contributors", "Forks", and so on centered. Also,
add a link to the documentation section of the README in the table of
contents.

* docs: fix broken LICENSE link in the README

The LICENSE link was broken in the README file, and has been fixed now.

* docs: format README file

Apply formatting to the README file, and make code lines shorter.

* docs: add min width to table columns

Add a minimum width to table columns to make them easier to read.

* docs: move examples to the README usage section

Move code examples from the documentation tables to the usage section,
since when the code examples are in the tables, they take up too much
space.

* docs: add better formatting to code examples

Make the code examples more organized.

* docs: add remaining README docs for classes

Add the remaining documentation for class methods.