feat: clean up code and add department functions (#26)
* feat: add concurrency limit and request interval

Add concurrency limit and request interval options
to the `limitConcurrency` function. That way, many
API requests can be made without overloading the
API.

* feat: try to get `limitConcurrency` to work

Work on trying to make
`apiConfig.concurrencyLimit` > 1 work with
`limitConcurrency`. Currently, there is some sort
of data race issue.

* build: remove `axios` as a dependency

Now, this package has *zero* runtime dependencies!
There is of course still the `package.json`
"devDependencies". Anyways, instead of `axios`,
`fetch` is simply used for web requests.

* build: change package name

Change the package name to not have "sfu" in it.

* docs: update README code examples

Update the README code examples to reflect the
current state of the package.

* feat: add department course wrapper functions

Add wrapper functions to get courses in a
department.

* feat: add `years` and `terms` wrapper functions

Add the `years` and `terms` wrapper functions.
Also, add tests for them.

* feat: add `Department` class

Add a `Department` class to be used in the
`departments` wrapper function.

Also, add a test case for using a for loop on a
`Course` to get its `CourseSections`.