feat: add wrapper test delay and fail fast for 404
Add a delay between wrapper tests and fail fast for 404 errors. This
way, errors from the API being spammed should not happen during tests.
Then, this wrapper code can be modified to fail immediately if a 404
error is received from the API. Such an error should very likely
actually be a 404 error, as long as the API is not being spammed.