feat: do not throw empty response error (#41)
This error can be thrown when there wasn't actually an API error. It is
possible for an empty response to be the "true" response.

This issue is because an empty array is used as a response for both
cases like: "that department exists and has no courses" and "that
department does not exist, and so no courses exist for this non-existing
department".

I think it is better to not throw the error though. A user can handle
requesting a department that doesn't exist themselves.